--- policyengine_canada/parameters/gov/provinces/ns/tax/income/income_assistance/eligibility/dependant_age_eligibility.yaml
@@ -1,9 +1,0 @@
-description: Nova Scotia qualifies dependents below the following age limit under the Income Assistance (IA) if they attend an approved educational program.
-values:
-  2021-01-01: 21
-metadata:
-  unit: year
-  label: Nova Scotia Income Assistance dependant age ceiling 
-  reference:
-  - title: Employment Support and Income Assistance Act, Section 21, Eligibility (8), (1)
-    href: https://novascotia.ca/just/regulations/regs/esiaregs.htm#TOC2_2
