--- policyengine_canada/parameters/gov/provinces/ns/tax/income/income_assistance/eligibility/lower_age_eligibility.yaml
@@ -1,9 +1,0 @@
-description: Nova Scotia qualifies fillers at or above the following age limit for the Income Assistance (IA) if they meet certain eligibility criteria.
-values:
-  2021-01-01: 16
-metadata:
-  unit: year
-  label: Nova Scotia Income Assistance lower age eligibility
-  reference:
-  - title: Employment Support and Income Assistance Act, Section 21, Eligibility (8), (1)
-    href: https://novascotia.ca/just/regulations/regs/esiaregs.htm#TOC2_2
